fix(amazonq): change the findings suffix, adjust the stored findings when lines are modified.
