+++ packages/core/src/codewhisperer/activation.ts
@@ -476,6 +476,7 @@ export async function activate(context: ExtContext): Promise<void> {
                               CodeWhispererConstants.amazonqIgnoreNextLine
                           )
                 )
+                SecurityIssueProvider.instance.cleanOldFiles()
             }),
             vscode.window.createTreeView(SecurityIssueTreeViewProvider.viewType, {
                 treeDataProvider: SecurityIssueTreeViewProvider.instance,
